HistoryDeepResearch/smolagents_project_root/examples/open_deep_research/requirements.txt
anthropic>=0.37.1
datasets>=2.21.0
huggingface_hub>=0.23.4
openai>=1.52.2
transformers>=4.46.0
torch>=2.2.2
torchvision>=0.17.2
scikit-learn
scipy
sympy
smolagents[litellm]
tqdm>=4.66.4
beautifulsoup4>=4.12.3
mammoth>=1.8.0
markdownify>=0.13.1
openpyxl
pypdf>=5.1.0
PyPDF2
pdfminer.six==20231228
python-pptx>=1.0.2
puremagic>=1.28
SpeechRecognition
youtube_transcript_api>=0.6.2
numpy>=2.1.2
pandas>=2.2.3
numexpr>=2.10.1
pydub>=0.25.1
xlrd
pathvalidate>=3.2.1
python-dotenv>=1.0.1
requests>=2.32.3
google-search-results==2.4.2 
easyocr>=1.7.1
translate>=3.6.1  
langchain-core
playwright
posthog   
langchain-openai
chess
pubchempy
Bio
audioop-lts<1.0; python_version >= "3.13"
